# backend/docs/openapi/OverfitterAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "Overfitter",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "localhost:4000"
    },
    {
      "url": "overfitter.jamkelley.com/api"
    }
  ],
  "paths": {
...
    "/accessories": {
      "get": {
        "summary": "Get All Accessory Items",
        "operationId": "getAllAccessoryItems",
        "tags": [
          "Accessories"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AccessoryResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "post": {
        "summary": "Create Accessory Item",
        "operationId": "createAccessoryItem",
        "tags": [
          "Accessories"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AccessoryBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AccessoryResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete All Accessory Items",
        "operationId": "deleteAllAccessoryItems",
        "tags": [
          "Accessories"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AccessoryResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/accessories/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/idParam"
        }
      ],
      "get": {
        "summary": "Get Accessory Item by Id",
        "operationId": "getAccessoryByIdItem",
        "tags": [
          "Accessories"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AccessoryResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete Accessory Item by Id",
        "operationId": "deleteAccessoryItemById",
        "tags": [
          "Accessories"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AccessoryResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "put": {
        "summary": "Update Accessory Item by Id",
        "operationId": "updateAccessoryItemById",
        "tags": [
          "Accessories"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AccessoryBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AccessoryResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/accessories/search": {
      "post": {
        "summary": "Search for Accessory Item",
        "operationId": "searchAccessoryItem",
        "tags": [
          "Accessories"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ClothingItemSearchBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AccessoryResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/bottoms": {
      "get": {
        "summary": "Get All Bottom Items",
        "operationId": "getAllBottomItems",
        "tags": [
          "Bottoms"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BottomResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "post": {
        "summary": "Create Bottom Item",
        "operationId": "createBottomItem",
        "tags": [
          "Bottoms"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/BottomBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/BottomResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete All Bottom Items",
        "operationId": "deleteAllBottomItems",
        "tags": [
          "Bottoms"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BottomResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/bottoms/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/idParam"
        }
      ],
      "get": {
        "summary": "Get Bottom Item by Id",
        "operationId": "getBottomByIdItem",
        "tags": [
          "Bottoms"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BottomResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete Bottom Item by Id",
        "operationId": "deleteBottomItemById",
        "tags": [
          "Bottoms"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BottomResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "put": {
        "summary": "Update Bottom Item by Id",
        "operationId": "updateBottomItemById",
        "tags": [
          "Bottoms"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/BottomBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/BottomResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/bottoms/search": {
      "post": {
        "summary": "Search for Bottom Item",
        "operationId": "searchBottomItem",
        "tags": [
          "Bottoms"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ClothingItemSearchBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/BottomResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "AccessoryResponse": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/WithId"
                        },
                        {
                          "$ref": "#/components/schemas/Accessory"
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "AccessoryResponseList": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/WithId"
                },
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/WithId"
                          },
                          {
                            "$ref": "#/components/schemas/Accessory"
                          }
                        ]
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "BottomResponse": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/WithId"
                        },
                        {
                          "$ref": "#/components/schemas/Bottom"
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "BottomResponseList": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/WithId"
                },
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/WithId"
                          },
                          {
                            "$ref": "#/components/schemas/Bottom"
                          }
                        ]
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorType"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorType"
            }
          }
        }
      },
      "UnexpctedError": {
        "description": "Unexpected error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorType"
            }
          }
        }
      }
    },
    "schemas": {
      "WithId": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      },
...
      "DatalessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "size_us_letter": {
        "type": "string",
        "enum": [
          "5XS",
          "4XS",
          "3XS",
          "2XS",
          "XS",
          "S",
          "M",
          "L",
          "XL",
          "2XL",
          "3XL",
          "4XL",
          "5XL",
          "UNKNOWN"
        ]
      },
      "Entity": {
        "type": "object",
        "properties": {
          "uriImage": {
            "type": "string",
            "pattern": "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)",
            "example": "https://www.pngmart.com/files/11/Rickrolling-PNG-Transparent-Image.png"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "timestampAddedRFC": {
            "type": "string",
            "format": "date-time"
          },
          "timestampLastModifiedRFC": {
            "type": "string",
            "format": "date-time"
          },
          "rating": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5
          }
        }
      },
      "ClothingItem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Entity"
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "sizeUSLetter": {
                "$ref": "#/components/schemas/size_us_letter"
              },
              "sizeUSNumber": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "brand": {
                "type": "string"
              },
              "timestampPurchasedRFC": {
                "type": "string",
                "format": "date-time"
              },
              "itemCondition": {
                "$ref": "#/components/schemas/item_condition"
              },
              "itemStatus": {
                "$ref": "#/components/schemas/item_status"
              },
              "numberOfWears": {
                "type": "integer",
                "minimum": 0,
                "maximum": 1000000
              },
              "wearsBeforeDirty": {
                "type": "integer",
                "minimum": 0,
                "maximum": 1000000
              },
              "wearsLeftBeforeDirty": {
                "type": "integer",
                "minimum": 0,
                "maximum": 1000000
              },
              "primaryColor": {
                "type": "string"
              },
              "secondaryColor": {
                "type": "string"
              },
              "accentColor": {
                "type": "string"
              },
              "pattern": {
                "type": "string"
              }
            }
          }
        ]
      },
      "accessory_type": {
        "type": "string",
        "enum": [
          "RING",
          "NECKLACE",
          "EARRINGS",
          "PURSE",
          "UNKNOWN"
        ]
      },
      "Accessory": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ClothingItem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "$ref": "#/components/schemas/accessory_type"
              }
            }
          }
        ]
      },
      "bottom_type": {
        "type": "string",
        "enum": [
          "SWEAT",
          "DRESS",
          "UNDER",
          "SLEEP",
          "WORKOUT",
          "JEANS",
          "KAKIS",
          "UNKNOWN"
        ]
      },
      "bottom_inseam_type": {
        "type": "string",
        "enum": [
          "SHORT",
          "REGULAR",
          "TALL",
          "UNKNOWN"
        ]
      },
      "Bottom": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ClothingItem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "$ref": "#/components/schemas/bottom_type"
              },
              "sizeHipInchesLow": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "sizeHipInchesHigh": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "bottomInseamType": {
                "$ref": "#/components/schemas/bottom_inseam_type"
              },
              "inseamInches": {
                "type": "number",
                "minimum": 0,
                "maximum": 50
              }
            }
          }
        ]
      },
...
      "ErrorType": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "errorMessage": {
            "type": "string"
          },
          "error": {
            "type": "object",
            "nullable": true
          },
          "detail": {
            "type": "object",
            "nullable": true
          }
        }
      }
    },
    "parameters": {
      "idParam": {
        "name": "id",
        "in": "path",
        "description": "ID of the Item",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "requestBodies": {
      "AccessoryBody": {
        "description": "A JSON object containing accessory info",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Accessory"
            }
          }
        }
      },
      "BottomBody": {
        "description": "A JSON object containing bottom info",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Bottom"
            }
          }
        }
      },
...
      "EntitySearchBody": {
        "description": "A JSON object containing search info for an Entity",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "timestampAddedBeginRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "timestampAddedEndRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "timestampModifiedBeginRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "timestampModifiedEndRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "rating": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "ClothingItemSearchBody": {
        "description": "A JSON object containing search info for a ClothingItem",
        "content": {
          "application/json": {
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/requestBodies/EntitySearchBody"
                },
                {
                  "type": "object",
                  "properties": {
                    "sizeUSLetter": {
                      "type": "string"
                    },
                    "sizeUSNumber": {
                      "type": "integer"
                    },
                    "brand": {
                      "type": "string"
                    },
                    "timestampPurchasedBeginRFC": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "timestampPurchasedEndRFC": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "itemCondition": {
                      "type": "string"
                    },
                    "itemStatus": {
                      "type": "string"
                    },
                    "numberOfWearsMin": {
                      "type": "integer"
                    },
                    "numberOfWearsMax": {
                      "type": "integer"
                    },
                    "wearsBeforeDirtyMin": {
                      "type": "integer"
                    },
                    "wearsBeforeDirtyMax": {
                      "type": "integer"
                    },
                    "wearsLeftBeforeDirtyMin": {
                      "type": "integer"
                    },
                    "wearsLeftBeforeDirtyMax": {
                      "type": "integer"
                    },
                    "primaryColor": {
                      "type": "string"
                    },
                    "secondaryColor": {
                      "type": "string"
                    },
                    "accentColor": {
                      "type": "string"
                    },
                    "pattern": {
                      "type": "string"
                    },
                    "itemType": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}